Separated substring extension into separate file.
