+++ Sources/VHDLParsing/String+VHDLMethods.swift
@@ -116,30 +116,6 @@ extension String {
         .joined(separator: "\n")
     }
 
-    // @usableFromInline var wordsAndOperators: [String]? {
-    //     let words = self.words
-    //     let operations = Set<String>.vhdlOperations.union(Set<String>(["(", ")"]))
-    //     return words.flatMap {
-    //         var subWords = [$0]
-    //         operations.forEach { op in
-    //             subWords = subWords.flatMap { subWord in
-    //                 let splitted = subWord.components(separatedBy: op)
-                    
-    //             }
-    //         }
-    //         return subWords
-    //     }
-    // }
-
-    // @usableFromInline func splittingInclusively(word: String) -> [String] {
-    //     var splitted: [String] = []
-    //     var startIndex = self.startIndex
-    //     var index = self.startIndex
-    //     while index < self.endIndex {
-    //         self[startIndex...].startIndex(word: word)
-    //     }
-    // }
-
     /// Find all expressions within self that exist within a set of brackets. The substrings returned may also
     /// contain substrings with brackets within them.
     @usableFromInline var subExpressions: [Substring]? {
@@ -393,277 +369,3 @@ extension String {
     }
 
 }
-
-/// Add `startIndex`.
-extension Substring {
-
-    @usableFromInline var topExpressions: [Substring]? {
-        var startIndex = self.startIndex
-        var currentIndex = self.startIndex
-        var subExpressions: [Substring] = []
-        while currentIndex < self.endIndex {
-            if self[currentIndex] == "(" {
-                subExpressions.append(self[startIndex...currentIndex])
-                guard let sub = self[currentIndex...].uptoBalancedBracket else {
-                    return nil
-                }
-                subExpressions.append(sub)
-                currentIndex = sub.endIndex
-                startIndex = sub.endIndex
-            }
-            currentIndex = self.index(after: currentIndex)
-        }
-        subExpressions.append(self[startIndex...])
-        return subExpressions
-    }
-
-    /// Return a string that exists within self that starts with an open bracket and ends with the balanced
-    /// closing bracket.
-    @usableFromInline var uptoBalancedBracket: Substring? {
-        self.upToBalancedElements(startsWith: "(", endsWith: ")")
-    }
-
-    /// Find a string that starts with a specified string and ends with a specified string including
-    /// substrings following the same pattern. For example, consider the string \"a(b(c)d)e\", starting with
-    /// \"(\" and ending with \")\". The result would be \"(b(c)d)\".
-    /// - Parameters:
-    ///   - startsWith: The begining delimiter for the substring.
-    ///   - endsWith: The ending delimiter for the substring.
-    /// - Returns: The substring that starts with `startsWith` and ends with `endsWith` including any
-    /// strings within that match the same pattern.
-    @usableFromInline
-    func upToBalancedElements(startsWith: String, endsWith: String) -> Substring? {
-        guard !startsWith.isEmpty, !endsWith.isEmpty else {
-            return nil
-        }
-        let startSize = startsWith.count
-        let endSize = endsWith.count
-        var startCount = 0
-        var hasStarted = false
-        var index = self.startIndex
-        var beginIndex: String.Index?
-        while index < self.endIndex {
-            guard hasStarted else {
-                guard
-                    let startIndex = self.startIndex(for: startsWith),
-                    let nextIndex = self.index(startIndex, offsetBy: startSize, limitedBy: self.endIndex)
-                else {
-                    return nil
-                }
-                beginIndex = startIndex
-                index = nextIndex
-                hasStarted = true
-                startCount += 1
-                continue
-            }
-            let data = self[index...]
-            guard let endIndex = data.startIndex(for: endsWith) else {
-                return nil
-            }
-            if let nextStartIndex = data.startIndex(for: startsWith) {
-                guard nextStartIndex > endIndex else {
-                    startCount += 1
-                    index = self.index(nextStartIndex, offsetBy: startSize)
-                    continue
-                }
-            }
-            let lastIndex = self.index(endIndex, offsetBy: endSize)
-            guard startCount <= 1 else {
-                startCount -= 1
-                index = lastIndex
-                continue
-            }
-            guard let beginIndex = beginIndex else {
-                return nil
-            }
-            return self[beginIndex..<lastIndex]
-        }
-        return nil
-    }
-
-    /// Return the starting index of a substring value within self.
-    /// - Parameter value: The substring to search for.
-    /// - Returns: The first index within self that matches the substring.
-    @usableFromInline
-    func startIndex(for value: String) -> String.Index? {
-        let size = value.count
-        guard !value.isEmpty, self.count >= size else {
-            return nil
-        }
-        let offset = size - 1
-        let startIndex = self.index(self.startIndex, offsetBy: offset)
-        for i in self[startIndex...].indices {
-            guard
-                let wordStart = self.index(i, offsetBy: -offset, limitedBy: self.startIndex),
-                self[wordStart...i].lowercased() == value.lowercased()
-            else {
-                continue
-            }
-            return wordStart
-        }
-        return nil
-    }
-
-    /// Find the start index for a word.
-    /// - Parameter word: The word to search for.
-    /// - Returns: The index if the word was found.
-    @usableFromInline
-    func startIndex(word: String) -> String.Index? {
-        guard !word.isEmpty, !self.isEmpty else {
-            return nil
-        }
-        var index = self.startIndex
-        while index < self.endIndex {
-            guard let startIndex = self[index...].startIndex(for: word) else {
-                return nil
-            }
-            guard let previousIndex = self.index(startIndex, offsetBy: -1, limitedBy: self.startIndex) else {
-                guard let endIndex = self.index(
-                    startIndex, offsetBy: word.count, limitedBy: self.index(before: self.endIndex)
-                ) else {
-                    return startIndex
-                }
-                guard CharacterSet.whitespacesAndNewlines.contains(self.unicodeScalars[endIndex]) else {
-                    index = self.index(startIndex, offsetBy: word.count)
-                    continue
-                }
-                return startIndex
-            }
-            guard let endIndex = self.index(
-                startIndex, offsetBy: word.count, limitedBy: self.index(before: self.endIndex)
-            ) else {
-                guard CharacterSet.whitespacesAndNewlines.contains(self.unicodeScalars[previousIndex]) else {
-                    index = self.index(startIndex, offsetBy: word.count)
-                    continue
-                }
-                return startIndex
-            }
-            guard
-                CharacterSet.whitespacesAndNewlines.contains(self.unicodeScalars[previousIndex]),
-                CharacterSet.whitespacesAndNewlines.contains(self.unicodeScalars[endIndex])
-            else {
-                index = self.index(startIndex, offsetBy: word.count)
-                continue
-            }
-            return startIndex
-        }
-        return nil
-    }
-
-}
-
-// indirect enum SubExpression {
-
-//     case binary(lhs: SubExpression, rhs: SubExpression, operation: Substring)
-
-//     case unary(operation: Substring, operand: SubExpression)
-
-//     case precedence(value: SubExpression)
-
-//     case variable(value: Substring)
-
-//     fileprivate init?(rawValue: Substring) {
-//         if rawValue.hasPrefix("("), rawValue.hasSuffix(")") {
-//             guard let newExpression = SubExpression(rawValue: rawValue.dropFirst().dropLast()) else {
-//                 return nil
-//             }
-//             self = .precedence(value: newExpression)
-//             return
-//         }
-//         let words = String(rawValue).words
-//         if words.count == 1 && !Set<String>.vhdlOperations.contains(words[0]) {
-//             self = .variable(value: rawValue)
-//             return
-//         }
-//         guard words.count > 1 else {
-//             return nil
-//         }
-//         let labelledExpressions = words.compactMap { SubExpressionLabel(word: $0) }
-//     }
-
-// }
-
-// private enum SubExpressionLabel {
-
-//     case multiplication
-
-//     case division
-
-//     case subtraction
-
-//     case addition
-
-//     case lessThan
-
-//     case lessThanEqual
-
-//     case greaterThan
-
-//     case greaterThanEqual
-
-//     case equal
-
-//     case notEqual
-
-//     case negation
-
-//     case and
-
-//     case or
-
-//     case nand
-
-//     case nor
-
-//     case xor
-
-//     case xnor
-
-//     case subExpression(value: String)
-
-//     init?(word: String) {
-//         let trimmedWord = word.trimmingCharacters(in: .whitespacesAndNewlines)
-//         guard trimmedWord.words.count == 1 else {
-//             return nil
-//         }
-//         switch trimmedWord.lowercased() {
-//         case "+":
-//             self = .addition
-//         case "-":
-//             self = .subtraction
-//         case "*":
-//             self = .multiplication
-//         case "/":
-//             self = .division
-//         case "<":
-//             self = .lessThan
-//         case "<=":
-//             self = .lessThanEqual
-//         case ">":
-//             self = .greaterThan
-//         case ">=":
-//             self = .greaterThanEqual
-//         case "=":
-//             self = .equal
-//         case "/=":
-//             self = .notEqual
-//         case "not":
-//             self = .negation
-//         case "and":
-//             self = .and
-//         case "or":
-//             self = .or
-//         case "nand":
-//             self = .nand
-//         case "nor":
-//             self = .nor
-//         case "xor":
-//             self = .xor
-//         case "xnor":
-//             self = .xnor
-//         default:
-//             self = .subExpression(value: word)
-//         }
-//     }
-
-// }
